Umstellung auf neue Dateistruktur
Neue Dateistruktur teilweise umgesetzt
