+++ LUA/ak/demo-anlagen/ampel/Ampeldemo-Grundmodelle-test.lua
@@ -1,3 +1,10 @@
+Zugname = "#PLATZHALTER"
+
+print("Lade AkEepFunctions ...")
+require 'ak.eep.AkEepFunktionen'
+-- endregion
+
+clearlog()
 --------------------------------------------------------------------
 -- Zeigt erweiterte Informationen waehrend der Initialisierung an --
 --------------------------------------------------------------------
@@ -7,7 +14,7 @@ AkStartMitDebug = false
 -- Zeigt erweiterte Informationen waehrend der erste Schitte an   --
 --------------------------------------------------------------------
 print("Lade Ampeldemo-Grundmodelle-main ...")
-require 'ak-demo-lua.ampel.Ampeldemo-Grundmodelle-main'
+require 'ak.demo-anlagen.ampel.Ampeldemo-Grundmodelle-main'
 
 --------------------------------------------------------------------
 -- Zeige erweiterte Informationen an                              --
@@ -24,17 +31,41 @@ AkKreuzung.showSchaltungAlsInfo = false
 --------------------------------------------------------------------
 -- AkKreuzung.zaehlerZuruecksetzen()
 
-[EEPLuaData]
-DS_101 = "f=2,p=Rot,w=3,"
-DS_102 = "f=0,p=Gruen,w=0,"
-DS_103 = "f=0,p=Gruen,w=1,"
-DS_104 = "f=1,p=Rot,w=2,"
-DS_105 = "f=1,p=Rot,w=3,"
-DS_106 = "f=0,p=Gruen,w=0,"
-DS_107 = "f=2,p=Rot,w=1,"
-DS_108 = "f=1,p=Rot,w=2,"
-DS_121 = "f=0,p=Rot,w=2,"
-DS_122 = "f=0,p=Rot,w=2,"
-DS_123 = "f=0,p=Rot,w=0,"
-DS_124 = "f=0,p=Rot,w=0,"
-DS_125 = "f=0,p=Gruen,w=1,"
+
+
+
+
+-------------------------------------------------------------------
+AkKreuzung.debug = true
+KpBetritt(k1_r8)
+KpBetritt(k1_r8)
+assert(k1_r8.fahrzeuge == 2, k1_r8.anzahlFahrzeuge)
+AkKreuzung.zaehlerZuruecksetzen()
+assert(k1_r8.fahrzeuge == 0)
+-------------------------------------------------------------------
+local function run()
+    EEPTime = EEPTime + 20
+    EEPMain()
+end
+
+local function k1Print()
+    list = {}
+    for k, v in pairs(k1:getSchaltungen()) do table.insert(list, k) end
+    table.sort(list, AkKreuzungsSchaltung.hoeherePrioAls)
+
+    for k, v in ipairs(list) do
+        print(k .. ": " .. v:getName() .. " - Prio: " .. v:getPrio())
+    end
+end
+
+for i = 1, 10 do
+    print("Betritt Block")
+    KpBetritt(k1_r8)
+    run()
+    run()
+    run()
+    run()
+    print("Verlasse Block")
+    KpVerlasse(k1_r8,true)
+    run()
+end
