+++ LUA/ak/text/AkAusgabe.lua
