+++ Resourcen/Anlagen/Andreas_Kreuz-Demo-Testen/Andreas_Kreuz-Lua-Testbeispiel.lua
@@ -1,0 +1,7 @@
+require 'ak.demo-anlagen.testen.Andreas_Kreuz-Lua-Testbeispiel-main'
+
+-- Der wahre Schatz liegt in 
+-- EEP\LUA\ak\demo-anlagen\testen\Andreas_Kreuz-Lua-Testbeispiel-test
+
+[EEPLuaData]
+DN_1 = 1.000000
