--- ak-demo-anlagen/testen/Lua-Test-vor-EEP.lua
@@ -1,3 +1,0 @@
-require 'ak-demo-lua.testen.Lua-Test-vor-EEP-main'
-[EEPLuaData]
-DN_1 = 1.000000
